roles do estudante
O estuante ja tem permissoes no sitema

+++ app/Http/Controllers/EstudanteController.php
@@ -6,6 +6,7 @@ namespace App\http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Estudante;
 use App\Models\Curso;
+use Illuminate\Support\Facades\Auth;
 
 
 
@@ -16,26 +17,48 @@ class  EstudanteController extends Controller{
 
 public function index(){
 
+ if (Auth::user()->can('view', Estudante::class)) {
+    $cursos = Curso::all();
 $estudantes = Estudante:: paginate(8);
-return view('/estudante.index',compact('estudantes'));
+return view('/estudante.index',['estudantes' => $estudantes]);
+}else {
+    return redirect()->back()->with('error', 'Você não tem permissão para visualizar os Estudantes.');
 }
+}
+
+
+
 
 
 public function update_view($id){
-    $estudante = Estudante:: find($id);
-    return view('/estudante/edit', compact('estudante'));
+    if (Auth::user()->can('update', Estudante::class)) {
+        $cursos = Curso::all();
+    $estudante = Estudante:: find($id); 
+    return view('/estudante/edit', compact('estudante','cursos'));
 }
+else {
+    return redirect()->back()->with('error', 'Você não tem permissão para editar este Estudante.');
+}
+}
+
+
 public function create()
 {
-    $cursos = Curso::all();
-    return view('estudante.create',compact('cursos'));
+   
+    if (Auth::user()->can('create', Estudante::class)){
+        $cursos = Curso::all();
+        return view('estudante.create',compact('cursos'));
+    } else {
+        return redirect()->back()->with('error', 'Você não tem permissão para criar um novo Estudante.');
+    }
 }
 
 public function saveEstudante(Request $request)
 {
+     
 
-    $estudante = new Estudante();
-
+     if (Auth::user()->can('create', Estudante::class)) {    
+        $estudante = new Estudante();
         $estudante->nome = $request->nome;
         $estudante->apelido = $request->apelido;
         $estudante->curso_id = $request->curso_id;
@@ -47,9 +70,19 @@ public function saveEstudante(Request $request)
 
         return redirect()->route('estudanteIndex')->with('mensagem', 'estudate Cadastrado com sucesso!');
 
+}else {
+    // O usuário não tem permissão, exibe uma mensagem de erro ou redireciona para outra página
+    return redirect()->back()->with('error', 'Você não tem permissão para criar um novo Estudante.');
+}
+
+
 }
 public function update(Request $request, $id){
 
+    if (Auth::user()->can('update', Estudante::class)) {
+    
+
+       
     $estudante =  Estudante:: find($id);
     $estudante->nome=$request->nome;
     $estudante->apelido=$request->apelido;
@@ -61,21 +94,34 @@ public function update(Request $request, $id){
     $estudante->save();
 
     return redirect()->route('estudanteIndex')->with('mensagem', 'Estudante Actualizado com sucesso!');
-}
+}else {
+    return redirect()->back()->with('error', 'Você não tem permissão para editar este Estudante.');
+}}
 
 
     public function visualizar_view($id){
+
+        if (Auth::user()->can('update',Estudante::class)) {
+      
         $estudante = Estudante :: find($id);
         return view('/estudante/view', compact('estudante'));
-
+        }else {
+            return redirect()->back()->with('error', 'Você não tem permissão para editar este Estudante.');
+        }
     }
 
     public function delete($id)
     {
+        if (Auth::user()->can('delete', Estudante::class)) {
         $estudante = estudante::find($id);
         $estudante->delete();
-
         return redirect()->route('estudanteIndex')->with('successDelete', 'Estudante Excluido com Sucesso!');
+    
+        } else {
+            return redirect()->back()->with('error', 'Você não tem permissão para apagar este Estudante.');
+        }
+    }
+    
     }
 
 
@@ -94,6 +140,6 @@ public function update(Request $request, $id){
 
 
 
-}
+
